www.search: rename to www.search_views
