+++ config/settings/base.py
@@ -108,7 +108,7 @@ INSTALLED_APPS = [
     "itou.www.geiq_assessments_views",
     "itou.www.home",
     "itou.www.prescribers_views",
-    "itou.www.search",
+    "itou.www.search_views",
     "itou.www.companies_views",
     "itou.www.signup",
     "itou.www.invitations_views",
